+++ rows.go
@@ -135,7 +135,7 @@ func (r otRows) Close() (err error) {
 
 	err = r.Rows.Close()
 	if err != nil {
-		recordSpanError(r.span, r.cfg.SpanOptions, err)
+		recordSpanError(r.span, r.cfg, err)
 	}
 
 	return
@@ -159,6 +159,6 @@ func (r otRows) beforeNext() {
 func (r otRows) afterNext(err error) {
 	// io.EOF is not an error. It is expected to happen during iteration.
 	if err != nil && !errors.Is(err, io.EOF) {
-		recordSpanError(r.span, r.cfg.SpanOptions, err)
+		recordSpanError(r.span, r.cfg, err)
 	}
 }