+++ stmt.go
@@ -60,7 +60,7 @@ func (s *otStmt) ExecContext(
 		ctx, span = createSpan(ctx, s.cfg, method, true, s.query, args)
 
 		defer span.End()
-		defer recordSpanErrorDeferred(span, s.cfg.SpanOptions, &err)
+		defer recordSpanErrorDeferred(span, s.cfg, &err)
 	}
 
 	if execer, ok := s.Stmt.(driver.StmtExecContext); ok {
@@ -99,7 +99,7 @@ func (s *otStmt) QueryContext(
 	if filterSpan(ctx, s.cfg.SpanOptions, method, s.query, args) {
 		queryCtx, span = createSpan(ctx, s.cfg, method, true, s.query, args)
 		defer span.End()
-		defer recordSpanErrorDeferred(span, s.cfg.SpanOptions, &err)
+		defer recordSpanErrorDeferred(span, s.cfg, &err)
 	} else {
 		queryCtx = ctx
 	}